Struct varweeks_millis::varweek_date_mod::VarweekDate

pub struct VarweekDate {
    year: u32,
    varweek: u32,
    day: u32,
}

varweek_date format is fixed and global:
4 digit year has the unit c for CE common era, one space
2 digit for varweek (very similar to week) from 01v to 53v, has unit v, one space
1 digit for varweek-day from 1d (monday) to 7d(sunday)
example: 2021c 53v 7d

Fields
year: u32varweek: u32day: u32
Implementations

impl VarweekDate

pub fn from_yvd(year: u32, varweek: u32, day: u32) -> Option<Self>

constructor from year, varweek, day returns None on error

Example

let varweeks = varweeks_millis::VarweekDate::from_yvd(2021,09,3).unwrap();
assert_eq!(varweeks.to_string(),"2021c 09v 3d");

pub fn from_year_ordinal_day(year: u32, ordinal_day: u32) -> Option<Self>

constructor from year, ordinal_day returns None on error

Example

let varweeks = varweeks_millis::VarweekDate::from_year_ordinal_day(2021,45).unwrap();
assert_eq!(varweeks.to_string(),"2021c 07v 3d");

pub fn from_naive_date(nd: NaiveDate) -> Option<Self>

constructor from NaiveDate

Example

let nd = chrono::NaiveDate::from_ymd(2021,02,28);
let varweeks = varweeks_millis::VarweekDate::from_naive_date(nd).unwrap();
assert_eq!(varweeks.to_string() ,"2021c 09v 3d");

pub fn from_ymd(year: u32, month: u32, day: u32) -> Option<Self>

constructor from year, month, day returns None on error

Example

let varweeks = varweeks_millis::VarweekDate::from_ymd(2021,02,28).unwrap();
assert_eq!(varweeks.to_string(),"2021c 09v 3d");

pub fn from_str(s: &str) -> Option<Self>

constructor from str in VarweekDate format returns None on error

Example

let varweeks = varweeks_millis::VarweekDate::from_str("2021c 09v 3d").unwrap();
assert_eq!(varweeks.to_string(),"2021c 09v 3d");

pub fn year(&self) -> u32

getter year u32

Example

let varweeks = varweeks_millis::VarweekDate::from_str("2021c 09v 3d").unwrap();
assert_eq!(varweeks.year(),2021);

pub fn varweek(&self) -> u32

getter varweek u32

Example

let varweeks = varweeks_millis::VarweekDate::from_str("2021c 09v 3d").unwrap();
assert_eq!(varweeks.varweek(),9);

pub fn day(&self) -> u32

getter day u32

Example

let varweeks = varweeks_millis::VarweekDate::from_str("2021c 09v 3d").unwrap();
assert_eq!(varweeks.day(),3);

pub fn to_naive_date(&self) -> Option<NaiveDate>

convert varweek_date to chrono NaiveDate returns None on error

Example

let vd = varweeks_millis::VarweekDate::from_ymd(2021,02,28).unwrap();
let nd = vd.to_naive_date().unwrap();
assert_eq!(nd, chrono::NaiveDate::from_ymd(2021,02,28));